requirements.txt
# FastAPI and server
fastapi==0.128.6
uvicorn[standard]==0.40.0
python-multipart==0.0.22
pydantic==2.12.5
pydantic-settings==2.12.0

# Database
sqlalchemy==2.0.46
asyncpg==0.31.0
alembic==1.18.3

# Google GenAI SDK for Nano Banana
google-genai==1.62.0
Pillow==12.1.0

# Utilities
python-dotenv==1.2.1
aiofiles==25.1.0
uuid-extension==0.2.0
